third_party/blink/web_tests/external/wpt/css/css-sizing/keyword-sizes-on-replaced-element-expected.txt

This is a testharness.js-based test.
Found 6 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] .test 59
  assert_equals: \n<canvas width="100" height="100" class="test min-height stretch indefinite" data-expected-width="10" data-expected-height="110"></canvas>\nheight expected 110 but got 10
[FAIL] .test 60
  assert_equals: \n<canvas width="100" height="100" class="test max-height stretch indefinite" data-expected-width="510" data-expected-height="110"></canvas>\nheight expected 110 but got 510
[FAIL] .test 62
  assert_equals: \n<canvas width="100" height="100" class="test min-height stretch indefinite" style="min-width: 50px" data-expected-width="60" data-expected-height="110"></canvas>\nheight expected 110 but got 10
[FAIL] .test 63
  assert_equals: \n<canvas width="100" height="100" class="test max-height stretch indefinite" style="max-width: 50px" data-expected-width="60" data-expected-height="60"></canvas>\nheight expected 60 but got 510
[FAIL] .test 65
  assert_equals: \n<canvas width="100" height="100" class="test min-height stretch indefinite" style="min-width: 150px" data-expected-width="160" data-expected-height="160"></canvas>\nheight expected 160 but got 10
[FAIL] .test 66
  assert_equals: \n<canvas width="100" height="100" class="test max-height stretch indefinite" style="max-width: 150px" data-expected-width="160" data-expected-height="110"></canvas>\nheight expected 110 but got 510
Harness: the test ran to completion.
